Homework_13/requirements.txt
Django==3.2
djangorestframework==3.12.4
Pygments==2.8.1
